# Conflicts:
#	src/main/java/uk/co/ogauthority/pwa/model/form/pwaapplications/views/techdrawings/PipelineDrawingSummaryView.java
#	src/main/java/uk/co/ogauthority/pwa/service/pwaapplications/shared/techdrawings/PadTechnicalDrawingService.java

+++ src/main/java/uk/co/ogauthority/pwa/service/fileupload/PadFileService.java
@@ -112,6 +112,7 @@ public class PadFileService {
    * Fully link temporary files that are still present, update file descriptions, delete files that have been deleted onscreen.
    * @param uploadForm containing files to update
    * @param pwaApplicationDetail we are updating files for
+   * @param updateMode the mode used when updating files
    * @param purpose of files being updated
    * @param user updating the files
    */
@@ -119,6 +120,7 @@ public class PadFileService {
   public void updateFiles(UploadMultipleFilesWithDescriptionForm uploadForm,
                           PwaApplicationDetail pwaApplicationDetail,
                           ApplicationFilePurpose purpose,
+                          FileUpdateMode updateMode,
                           WebUserAccount user) {
 
     Map<String, UploadFileWithDescriptionForm> uploadedFileIdToFormMap = getFileIdToFormMap(uploadForm);
@@ -144,41 +146,10 @@ public class PadFileService {
     });
 
     padFileRepository.saveAll(filesToUpdate);
-    deleteAppFileLinksAndUploadedFiles(filesToRemove, user);
 
-  }
-
-  /**
-   * Fully link temporary files that are still present, update file descriptions, delete files that have been deleted onscreen.
-   * @param uploadForm containing files to update
-   * @param pwaApplicationDetail we are updating files for
-   * @param purpose of files being updated
-   * @param user updating the files
-   */
-  @Transactional
-  public void makeFilesPermanent(UploadMultipleFilesWithDescriptionForm uploadForm,
-                          PwaApplicationDetail pwaApplicationDetail,
-                          ApplicationFilePurpose purpose,
-                          WebUserAccount user) {
-
-    Map<String, UploadFileWithDescriptionForm> uploadedFileIdToFormMap = getFileIdToFormMap(uploadForm);
-
-    var existingLinkedFiles = padFileRepository.findAllByPwaApplicationDetailAndPurpose(pwaApplicationDetail, purpose);
-
-    var filesToUpdate = new HashSet<PadFile>();
-
-    // if file is still in list of uploaded files, update description and add to update set
-    // else file can be deleted so add to remove set
-    existingLinkedFiles.forEach(existingFile -> {
-
-      if (uploadedFileIdToFormMap.containsKey(existingFile.getFileId())) {
-        updateFileDescriptionAndFullyLink(existingFile, uploadedFileIdToFormMap.get(existingFile.getFileId()));
-        filesToUpdate.add(existingFile);
-      }
-
-    });
-
-    padFileRepository.saveAll(filesToUpdate);
+    if (updateMode == FileUpdateMode.DELETE_UNLINKED_FILES) {
+      deleteAppFileLinksAndUploadedFiles(filesToRemove, user);
+    }
 
   }
